+++ src/backend/interfaces/REST/go.server/cmd/caliopen_rest/cli_cmds/root.go
@@ -20,7 +20,7 @@ var (
 	}
 )
 
-const __version__ = "0.5.2"
+const __version__ = "0.5.4"
 
 func init() {
 	cobra.OnInitialize()